Refactor: seeders and utils class.
- Refactorized the DbInitializer class and separated the seeders for Roles and Users to individual classes.
- Added the Utils class for common methods.
